Localise number on Y axis on the charts
All numbers on the page are localised, with the exception of numbers
on the Y axis on the charts. Where I am, this means that numbers on
the page are formatted with commas for thousands "93,873", but on
the Y axis of the chart it has "80000". This commit localises the
number of the Y axis, so it now shows "80,000".