Change how React mounts into its DOM container
Previously, the app mounted to the `#root` element. The meant that we
couldn't also _really_ use it for back-to-top navigation as the fragment
in the url name would look weird. Instead, this renames the containing
element to `#top` for navigation purposes and we instruct React to mount
via a class based querySelector on the new `.app-root` class.

Issue: PublicHealthEngland/coronavirus-dashboard#97