Finish back-to-top links' look and feel
This includes making a small, internal-only, component to hold the svg
content for the up arrow and improving the logic on how the overlay and
inline modes made themselves visible.

The back-to-top link SVG, HTML & CSS comes from
https://github.com/alphagov/govuk-design-system/blob/master/views/partials/_back-to-top.njk
and
https://github.com/alphagov/govuk-design-system/blob/master/src/stylesheets/components/_back-to-top.scss
but doesn't appear to be in the `govuk-react-jsx` package, so I included
it directly for simplicity.

Issue: PublicHealthEngland/coronavirus-dashboard#97