Add aria-label to the map canvas
It looks like the default aria-label ("Map") is coming directly from mapbox-gl:

https://github.com/mapbox/mapbox-gl-js/blame/d0f42e4274a0972f27c1e7f1546aea58491d2a4c/src/ui/map.js#L2285

There doesn't seem to be a nice way to configure this as a user of
mapbox-gl, so the easiest thing to do is just fish the canvas out of the
DOM and reset the attribute.

Note that this needs to be done after `L.map()` is called, because the
`mapbox` doesn't get its DOM until its used in the Leaflet map.

Fixes #80