On header link  redirect to / route rather than hardcoded staging site
This commit is a fix for #251.
On the master site https://coronavirus.data.gov.uk/ the main header
redirects to https://coronavirus-staging.data.gov.uk/

On the staging site https://coronavirus-staging.data.gov.uk/ the main
header redirects to https://coronavirus.data.gov.uk/

I'm assuming this is not intentional and is a bug, and set the header
link to redirect to / route, so that on whatever site this project is
deployed, it will always redirect to the base index page.